requirements.txt
asgiref==3.2.10
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
Django==3.1.2
django-cors-headers==3.5.0
djangorestframework==3.12.1
google-api-core==1.23.0
google-auth==1.22.1
google-cloud-pubsub==2.1.0
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.33.1
idna==2.10
libcst==0.3.13
mypy-extensions==0.4.3
Pillow==8.0.1
proto-plus==1.11.0
protobuf==3.13.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.11
